+++ packages/did-ui-react/src/components/config-provider/utils.ts
@@ -1,5 +1,7 @@
 import ConfigProvider from './';
 import { dealURLLastChar, did } from '../../utils';
+import { NetworkType, TCustomNetworkType } from '../../types';
+import { Portkey_Bot_Webapp } from '../../constants/telegram';
 
 export const getServiceUrl = () => {
   if (!ConfigProvider.config.serviceUrl) throw Error('Please config serviceUrl');
@@ -23,4 +25,14 @@ export const getStorageInstance = () => {
   const storageMethod = ConfigProvider.config.storageMethod || did.config.storageMethod;
   if (!storageMethod) throw Error('Please config storageMethod');
   return storageMethod;
+};
+
+export const getDappTelegramLink = () => {
+  const dappTelegramLink = ConfigProvider.getConfig('dappTelegramLink') as string;
+  if (!dappTelegramLink) throw Error('Please set dappTelegramLink in GlobalConfig');
+  return dappTelegramLink;
+};
+
+export const getPortkeyBotWebappLink = (ctw: TCustomNetworkType, network?: NetworkType) => {
+  return network ? Portkey_Bot_Webapp[ctw][network] : Portkey_Bot_Webapp[ctw].MAINNET;
 };